+++ internal/core/domain/command/image.go
@@ -34,7 +34,7 @@ func (i *Image) GetCommand() string {
 
 func (i *Image) Respond(ctx context.Context, timeout time.Duration, message *domain.Message) error {
 	l := log.With().
-		Int("messageId", message.ID).
+		Int64("messageId", message.ID).
 		Int64("chatId", message.ChatID).
 		Str("imageURL", message.ImageURL).
 		Str("command", i.GetCommand()).
